deprecate Prepper class
