+++ data-prepper-api/src/test/java/com/amazon/dataprepper/model/processor/AbstractProcessorTest.java
@@ -1,50 +1,56 @@
 /*
- * Copyright OpenSearch Contributors
- * SPDX-License-Identifier: Apache-2.0
+ *  SPDX-License-Identifier: Apache-2.0
+ *
+ *  The OpenSearch Contributors require contributions made to
+ *  this file be licensed under the Apache-2.0 license or a
+ *  compatible open source license.
+ *
+ *  Modifications Copyright OpenSearch Contributors. See
+ *  GitHub history for details.
  */
 
-package com.amazon.dataprepper.model.prepper;
+package com.amazon.dataprepper.model.processor;
 
-import com.amazon.dataprepper.model.configuration.PluginSetting;
 import com.amazon.dataprepper.metrics.MetricNames;
 import com.amazon.dataprepper.metrics.MetricsTestUtil;
+import com.amazon.dataprepper.model.configuration.PluginSetting;
 import com.amazon.dataprepper.model.record.Record;
+import io.micrometer.core.instrument.Measurement;
+import io.micrometer.core.instrument.Statistic;
+import org.junit.Assert;
+import org.junit.Test;
+
 import java.util.Arrays;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.List;
 import java.util.StringJoiner;
 import java.util.stream.Collectors;
-import io.micrometer.core.instrument.Measurement;
-import io.micrometer.core.instrument.Statistic;
-import org.junit.Assert;
-import org.junit.Test;
 
-@Deprecated
-public class AbstractPrepperTest {
+class AbstractProcessorTest {
 
     @Test
     public void testMetrics() {
-        final String prepperName = "testPrepper";
+        final String processorName = "testProcessor";
         final String pipelineName = "pipelineName";
         MetricsTestUtil.initMetrics();
 
-        PluginSetting pluginSetting = new PluginSetting(prepperName, Collections.emptyMap());
+        PluginSetting pluginSetting = new PluginSetting(processorName, Collections.emptyMap());
         pluginSetting.setPipelineName(pipelineName);
-        AbstractPrepper<Record<String>, Record<String>> prepper = new PrepperImpl(pluginSetting);
+        AbstractProcessor<Record<String>, Record<String>> processor = new ProcessorImpl(pluginSetting);
 
-        prepper.execute(Arrays.asList(
+        processor.execute(Arrays.asList(
                 new Record<>("Value1"),
                 new Record<>("Value2"),
                 new Record<>("Value3")
         ));
 
         final List<Measurement> recordsInMeasurements = MetricsTestUtil.getMeasurementList(
-                new StringJoiner(MetricNames.DELIMITER).add(pipelineName).add(prepperName).add(MetricNames.RECORDS_IN).toString());
+                new StringJoiner(MetricNames.DELIMITER).add(pipelineName).add(processorName).add(MetricNames.RECORDS_IN).toString());
         final List<Measurement> recordsOutMeasurements = MetricsTestUtil.getMeasurementList(
-                new StringJoiner(MetricNames.DELIMITER).add(pipelineName).add(prepperName).add(MetricNames.RECORDS_OUT).toString());
+                new StringJoiner(MetricNames.DELIMITER).add(pipelineName).add(processorName).add(MetricNames.RECORDS_OUT).toString());
         final List<Measurement> elapsedTimeMeasurements = MetricsTestUtil.getMeasurementList(
-                new StringJoiner(MetricNames.DELIMITER).add(pipelineName).add(prepperName).add(MetricNames.TIME_ELAPSED).toString());
+                new StringJoiner(MetricNames.DELIMITER).add(pipelineName).add(processorName).add(MetricNames.TIME_ELAPSED).toString());
 
         Assert.assertEquals(1, recordsInMeasurements.size());
         Assert.assertEquals(3.0, recordsInMeasurements.get(0).getValue(), 0);
@@ -58,8 +64,8 @@ public class AbstractPrepperTest {
                 0.2));
     }
 
-    public static class PrepperImpl extends AbstractPrepper<Record<String>, Record<String>> {
-        public PrepperImpl(PluginSetting pluginSetting) {
+    public static class ProcessorImpl extends AbstractProcessor<Record<String>, Record<String>> {
+        public ProcessorImpl(PluginSetting pluginSetting) {
             super(pluginSetting);
         }
